+++ app/ExternalFiles/Synchronization.php
@@ -434,11 +434,14 @@ class Synchronization {
 		add_action( 'eml_after_file_save', array( $this, 'mark_as_synced' ), 10, 3 );
 
 		// add counter handling.
-		add_action( 'eml_after_file_save', array( $this, 'update_url_count' ), 10, 0 );
 		add_action( 'eml_file_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
+		add_action( 'eml_file_directory_import_file_check', array( $this, 'update_url_count' ), 10, 0 );
 		add_action( 'eml_ftp_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
+		add_action( 'eml_ftp_directory_import_file_check', array( $this, 'update_url_count' ), 10, 0 );
 		add_action( 'eml_http_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
+		add_action( 'eml_http_directory_import_file_check', array( $this, 'update_url_count' ), 10, 0 );
 		add_action( 'eml_sftp_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
+		add_action( 'eml_sftp_directory_import_file_check', array( $this, 'set_url_max_count' ), 10, 2 );
 		add_action( 'eml_before_file_list', array( $this, 'change_process_title' ) );
 
 		// update the sync title on each file.